Keep the KDA checkpoint slot index 64-bit inside the store kernel
`_store_cache_checkpoints_kernel` multiplies the loaded 32-bit slot index by
the per-slot element strides of the convolution and recurrent state pools.
Loading it as int64 keeps the products 64-bit, so a pool large enough for
`slot * stride` to exceed 2^31 elements addresses the right slot instead of
wrapping; the stored values and every other index are unchanged.

Compatibility: no behavior change for pools below the wrap point (the served
TP9 configuration; overflow needs more than about 10,900 recurrent slots at
the 12-head, 128 x 128 fp32 per-rank stride). Matches upstream
vllm-project/vllm #55924.

Validation: source-level only in this commit; the kernel is a Triton JIT
function exercised by the existing KDA prefill checkpoint tests on a GPU.

+++ vllm/models/kimi_k3/nvidia/kda.py
@@ -335,7 +335,9 @@ def _store_cache_checkpoints_kernel(
 ):
     seq_idx = tl.program_id(0)
     cols = tl.program_id(1) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
-    state_idx = tl.load(checkpoint_state_indices_ptr + seq_idx)
+    # The slot index multiplies a per-slot element stride below; keep the
+    # product in 64 bits so large state pools cannot wrap a 32-bit offset.
+    state_idx = tl.load(checkpoint_state_indices_ptr + seq_idx).to(tl.int64)
     checkpoint_offset = tl.load(
         checkpoint_offsets_ptr + seq_idx * checkpoint_offset_stride
     )
